+++ protected/modules/yupe/install/yupe.php
@@ -75,11 +75,11 @@ return [
 ]
     ,
     'commandMap' => [
-    'yupe' => [
-        'class' => 'application.modules.yupe.commands.YupeCommand',
-    ],
-    'testenv' => [
-        'class' => 'application.modules.yupe.commands.TestEnvCommand'
+        'yupe' => [
+            'class' => 'application.modules.yupe.commands.YupeCommand',
+        ],
+        'testenv' => [
+            'class' => 'application.modules.yupe.commands.TestEnvCommand'
+        ]
     ]
-]
 ];